Made Chat and Comment Components, for factorization
